V1.0.0.3: 修复了IApplyResponseArgs.EventId属性类型错误的问题。(int => long) 将IApplyResponseArgs, INewApplyEventArgs, NewApplyEventArgs独立出一个文件。 添加MiraiHttpSession.UnknownMessageEvt事件, 每当Session收到未知消息时将引发它。

+++ Mirai-CSharp/Models/EventArgs/Group/GroupApplyEventArgs.cs
@@ -18,7 +18,7 @@ namespace Mirai_CSharp.Models
 
         }
 
-        public GroupApplyEventArgs(string fromGroupName, int eventId, long fromGroup, long fromQQ, string nickName) : base(eventId, fromGroup, fromQQ, nickName)
+        public GroupApplyEventArgs(string fromGroupName, long eventId, long fromGroup, long fromQQ, string nickName) : base(eventId, fromGroup, fromQQ, nickName)
         {
             FromGroupName = fromGroupName;
         }
